refactor: extract command idempotency types
